[chore] parseErrorResult 분리 및 적용
